requirements.txt
deepdiff==6.7.1
idna==3.6
ordered-set==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
PyYAML==6.0.1
requests==2.31.0
typing_extensions==4.7.1
urllib3==1.26.7
